+++ fastcrypto/src/lib.rs
@@ -84,10 +84,6 @@ pub mod utils_tests;
 #[path = "tests/secp256r1_group_tests.rs"]
 pub mod secp256r1_group_tests;
 
-#[cfg(test)]
-#[path = "tests/bn254_group_tests.rs"]
-pub mod bn254_group_tests;
-
 pub mod traits;
 
 #[cfg(feature = "aes")]
